requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
pandas
scikit-learn
faiss-cpu
python-multipart
pydantic
streamlit